Add files via upload
In this project, we aim to fully simulate a digital communication system and observe the impact of various parameters on its performance. The project entails the implementation of functions for signal division, pulse shaping, analog modulation and demodulation, channel simulation, and matched filtering. We focus on binary sequences to approximate a time-real system and analyze the system’s behavior in the presence of AWGN. The project provides insights into the practical aspects of communication systems engineering & signal processing.